timing and threading in tests
When running later you would need to give some time for the variable tray to come up.  I put in 1 second you might put more,  After I was able to access the 'tray' the first test failed because it was on the wrong thread.  This passes for me and I can then build